backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.1.2
gunicorn==21.2.0

